// FrontEndBase_Aulas_0043/src/ImportBackend/Interfaces/ConsultaPedidoERPInterfaces.ts
export interface ConsultaERP_PedidoInterface {
  identificador: number;
  nrPedidoCliente: string;
  observacao: string;
  observacaoUsoInterno: string;
  percFrete: number;
  valorFrete: number;
  percDesconto: number;
  valorDesconto: number;
  valorTotalBruto: number;
  valorTotalComImpostos: number;
  dataAtualizacao: number;
  valorTotal: number;
  percSeguro: number;
  valorSeguro: number;
  percDespAcessoria: number;
  valorDespAcessoria: number;
  condPagMut?: any;
  percDescFinanceiro?: any;
  percComissao: number;
  representante: string;
  representanteIdentificador: number;
  indicanteIdentificador?: any;
  condicoesPagamento: string;
  condicoesPagamentoIdentificador: number;
  tipoFrete: string;
  tipoFreteIdentificador: number;
  liberacaoPedidoPedIdentificador?: any;
  transportador: string;
  transportadorIdentificador: number;
  dataPrevisaoSaida: number;
  dataPrevisaoFat: number;
  naturezaOperacao: string;
  naturezaOperacaoIdentificador: number;
  unidadeFatCliente: string;
  unidadeFatClienteIdentificador: number;
  clienteIdentificador: number;
  dataEmissao: number;
  dataCadastro: number;
  empresa: string;
  empresaIdentificador: number;
  itemPedido: ItemPedido[];
  tipoDesconto: number;
  observacaoSistema?: any;
  tipoFreteInf: number;
  tipoSeguroInf: number;
  tipoDespAcessInf: number;
  destacarFrete: number;
  destacarSeguro: number;
  destacarDesconto: number;
  destacarDespAcessoria: number;
  informarLocalEntrega?: any;
  enderecoEntrega?: any;
  percFreteInf: number;
  valorFreteInf: number;
  percDescontoInf: number;
  valorDescontoInf: number;
  percSeguroInf: number;
  valorSeguroInf: number;
  percDespAcessoriaInf: number;
  valorDespAcessoriaInf: number;
  usuario?: any;
  usuarioIdentificador?: any;
  tipoConsumidor?: any;
  diasMediosCondPag: number;
  infPagamentoPedido: InfPagamentoPedido[];
  serialForSync: string;
  nfceConsumidor?: any;
  situacaoPedidoIdentificador: number;
  situacaoPedido: string;
  pessoaAutorizada?: any;
  pessoaAutorizadaIdentificador?: any;
  meioPagamento: string;
  meioPagamentoIdentificador: number;
  nrTotalItens: number;
  qtdeTotalItens: number;
  vlrTotalCusto: number;
  identificadorERP?: any;
  reservarEstoque?: any;
  modoReservaEstoque: number;
  centroEstoqueReserva?: any;
  centroEstoqueReservaIdentificador?: any;
  moeda: string;
  moedaIdentificador: number;
}
interface InfPagamentoPedido {
  identificador: number;
  meioPagamento: string;
  meioPagamentoIdentificador: number;
  valor: number;
  nrAutorizacao?: any;
  titulos: Titulo[];
  tipoPagamentoNFe?: any;
  tipoPagamentoNFeIdentificador?: any;
  bandeiraCartaoNFe?: any;
  bandeiraCartaoNFeIdentificador?: any;
  credenciadoraCreditoDebito?: any;
  credenciadoraCreditoDebitoIdentificador?: any;
  condicoesPagamentoIdentificador: number;
  condicoesPagamento: string;
  tipoIntegracao?: any;
  valorTroco?: any;
  valorLiquido: number;
  dataFinanceiraCartao?: any;
  valorAcrescimo: number;
  valorDesconto: number;
  nrDiasMedios: number;
  parcelas?: any;
  naoGerarComissao?: any;
  tipoPagST?: any;
}
// ...
interface ItemPedido {
  identificador: number;
  dataAtualizacao: number;
  valorUnitario: number;
  valorFrete: number;
  valorDesconto: number;
  valorTotalBruto: number;
  valorTotal: number;
  percComissao: number;
  percDesconto: number;
  percFrete: number;
  percDespesaAcessoria: number;
  percSeguro: number;
  valorDespesaAcessoria: number;
  valorSeguro: number;
  valorSugerido: number;
  valorMinimo: number;
  valorMaximo: number;
  valorCusto: number;
  modeloFiscal: string;
  modeloFiscalIdentificador: number;
  produto: string;
  produtoIdentificador: number;
  produtoNome: string;
  produtoCodigoAuxiliar: string;
  unidadeMedida: string;
  unidadeMedidaIdentificador: number;
  quantidadeTotal: number;
  gradeItemPedido: GradeItemPedido[];
  fatorConversao: number;
  infoAdicionalItem: string;
  infoAdicionalItemAux: string;
  nrItemPedido: string;
  nrSequencial: number;
  centroEstoque?: any;
  centroEstoqueIdentificador?: any;
  descontoItem: number;
  despAcessItem: number;
  seguroItem: number;
  freteItem: number;
  tipoDesconto: number;
  tipoFrete: number;
  tipoSeguro: number;
  tipoDespAcessoria: number;
  nrPedido: string;
  tipoTabPreco: number;
  formTabDinCalcComissao: string;
  formTabDinCalcComissaoIdentificador: number;
  formTabDinCalcPreco: string;
  formTabDinCalcPrecoIdentificador: number;
  errors: any[];
  itemPedidoFiscal: ItemPedidoFiscal;
}
interface ItemPedidoFiscal {
  identificador: number;
  itemPedidoIdentificador: number;
  itemPedido: string;
  aliquotaIcms: number;
  vrIcmsTributado: number;
  vrIcmsIsento: number;
  vrIcmsOutros: number;
  vrIcms: number;
  vrIcmsSemAprov: number;
  vrDifAliquota: number;
  vrIpiTributado: number;
  vrIpiIsento: number;
  vrIpiOutros: number;
  vrIpiIndustria: number;
  vrIpiComercio: number;
  vrIpiObservacao: number;
  vrBcCalculoIcmsSt: number;
  vrBcCalculoIcms: number;
  vrIcmsSt: number;
  vrIrrf: number;
  vrIss: number;
  vrInss: number;
  vrInssNaoRetido: number;
  vrCofins: number;
  vrLei10833: number;
  vrOutros: number;
  vrFunrural: number;
  aliquotaIpi: number;
  indPeriodoApuracao: number;
  vrDespAduaneira: number;
  vrImpostoImportacao: number;
  vrBCImpostoImp: number;
  vrIof: number;
  vrCofinsSt: number;
  vrPisSt: number;
  vrPis: number;
  percReducaoBCIcms: number;
  aliquotaIcmsST: number;
  descontoPadraoIcmsST: number;
  indiceAlteracaoIcmsST: number;
  vrContSoc: number;
  aliquotaFunrural: number;
  aliquotaIss: number;
  aliquotaContSoc: number;
  aliquotaLei10833: number;
  aliquotaInss: number;
  aliquotaPis: number;
  aliquotaCofins: number;
  aliquotaOutros: number;
  aliquotaIrrf: number;
  aliquotaSestSenat: number;
  percRedBcInss: number;
  percRedSestSenat: number;
  valorSestSenat: number;
  calcularIcmsST: number;
  percRedFunrural: number;
  percRedIrrf: number;
  percRedOutros: number;
  percRedLei10833: number;
  percRedContSoc: number;
  vrfreteIcmsST?: any;
  vrNaoTribIcms?: any;
  baseCalcIcmsFreteST: number;
  vrCustoICMSST: number;
  vrBCCustoICMSST: number;
  vrDevICMSST: number;
  vrBCDevICMSST: number;
  vrICMSSTRet: number;
  vrBCICMSSTRet: number;
  vrICMSDispensado: number;
  aliquotaICMSSimples: number;
  valorICMSSimples: number;
  vrBCPis: number;
  vrBCCofins: number;
  aliquotaCIDE: number;
  quantidadeBCCIDE: number;
  valorCIDE: number;
  qtdCombTempAmb: number;
  ufConsumoCombIdentificador?: any;
  ufConsumoComb?: any;
  incidenciaPisCofinsIdentificador: number;
  incidenciaPisCofins: string;
  incidenciaIcmsIdentificador: number;
  incidenciaIcms: string;
  incidenciaIpiIdentificador: number;
  incidenciaIpi: string;
  ufIcmsStIdentificador?: any;
  ufIcmsSt?: any;
  modalidadeIcmsStIdentificador?: any;
  modalidadeIcmsSt?: any;
  modalidadeIcmsIdentificador: number;
  modalidadeIcms: string;
  categoriaStIdentificador?: any;
  categoriaSt?: any;
  valorBCFCP: number;
  aliquotaFCP: number;
  valorFCP: number;
  valorBCFCPSt: number;
  aliquotaFCPSt: number;
  valorFCPSt: number;
  valorBCFCPStRetido: number;
  aliquotaFCPStRetido: number;
  valorFCPStRetido: number;
  percentualDiferimento: number;
  valorIcmsDiferimento: number;
  aliquotaPisSt: number;
  aliquotaCofinsSt: number;
  vrBCPisSt: number;
  vrBCCofinsSt: number;
  valorBcIcmsUfDest: number;
  aliquotaInternaUFDest: number;
  aliquotaInterestadual: number;
  percPartilhaIcms: number;
  valorIcmsPartilhaDest: number;
  valorIcmsPartilhaRem: number;
  aliquotaFundoPobreza: number;
  valorFundoPobreza: number;
}
interface GradeItemPedido {
  identificador: number;
  gradeCorIdentificador: number;
  gradeCor: string;
  quantidade: number;
  reservarEstoque: number;
}